Create AmazonButton
Re-purpose Amazon Buttons to become useful for home automation.

Given a pre-known set of Amazon Buttons, this code will record button presses to a Google sheet.
Additional actions may be taken for each button.

By example, one of the buttons is being used as a doorbell. This could be combined with taking a photo of who is at the door and sending it to a phone app.

Still thinking on what to use the other buttons for... ideas?